python/parakeet-cuda-uv/pyproject.toml
[project]
name = "parakeet-cuda-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cuda-python>=12.3",
    "flask>=3.1.2",
    "keyboard>=0.13.5",
    "nemo-toolkit[asr]<=2.5.0",
    "numpy>=1.26.4",
    "pyperclip>=1.11.0",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
]
